New Structure
The Last failed attempt was because I rushed all the code and the cpp and header files werent compiling correctly so I decided to start fresh with a more general outlook and piece it together one at a time while testing each component as I go. This Time Its just the Account Class making sure everything goes right and so far its a success!